+++ samples/helloskinningbuffer_morebones.cpp
@@ -217,7 +217,7 @@ int main(int argc, char** argv) {
         mat4f translate[] = {mat4f::translation(float3(s, c, 0))};
 
         mat4f trans[9] = {};
-        for (uint i = 0; i < 9; i++) {
+        for (size_t i = 0; i < 9; i++) {
             trans[i] = filament::math::mat4f(1);
         }
         s *= 8;
@@ -231,7 +231,7 @@ int main(int argc, char** argv) {
             mat4f::translation(float3(0, -s, 0)),
             mat4f::translation(float3(s, -s, 0)),
             filament::math::mat4f(1)};
-        uint offset = ((uint)now) % 8;
+        size_t offset = ((size_t)now) % 8;
         trans[offset] = transA[offset];
 
         // Set transformation of the first bone
